pyproject.toml
[project]
name = "lab1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.7.0",
    "langchain-core>=1.0.5",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "lyricsgenius>=3.7.5",
    "pre-commit>=4.4.0",
    "pydantic-settings>=2.12.0",
    "streamlit>=1.41.1",
    "httpx>=0.28.1",
    "gradio>=4.44.0",
    "langchain-deepseek>=1.0.1",
]
